# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/admin/delete": {
            "get": {
                "description": "delete file",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "delete file",
                "parameters": [
                    {
                        "type": "string",
                        "description": "filename",
                        "name": "filename",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "converted",
                            "upload"
                        ],
                        "type": "string",
                        "description": "保存種類。指定無しの場合`converted`として動作。",
                        "name": "kind",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/convert": {
            "post": {
                "description": "convert file",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "convert file",
                "parameters": [
                    {
                        "enum": [
                            "pdf",
                            "png",
                            "gif",
                            "png8",
                            "webp",
                            "bmp",
                            "jpeg",
                            "jpg",
                            "tiff"
                        ],
                        "type": "string",
                        "description": "出力ファイルフォーマット",
                        "name": "format",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "入力ファイル",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "dpi",
                        "name": "dpi",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "width",
                        "name": "width",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "height",
                        "name": "height",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "x",
                        "name": "x",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "y",
                        "name": "y",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "resize",
                            "crop"
                        ],
                        "type": "string",
                        "description": "`width`もしくは`height`を指定した際の変形方法。指定無しの場合は`resize`として動作。",
                        "name": "transformMode",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/download": {
            "get": {
                "description": "download file",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "download file",
                "parameters": [
                    {
                        "type": "string",
                        "description": "filename",
                        "name": "filename",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/files": {
            "get": {
                "description": "get file list",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "get file list",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/ping": {
            "get": {
                "description": "do ping",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "ping",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}